+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Improve compatibility with some plugins fixing some archiving issues
 * Improve tracking settings for multi sites
 * Show a blog selector in the reporting page when using multi site
+* Fix super admins weren't always synced in multisite mode
+* Improve system report
 * Some minor fixes
 
 = 1.0.6 =